LayoutTests/imported/w3c/web-platform-tests/html/semantics/popovers/popover-stacking-expected.txt

Direct DOM children

Grandchildren

popovertoggletarget attribute relationship

Ancestor popover

Button
nested popovertoggletarget attribute relationship

Ancestor popover

Button
anchor attribute relationship

Child popover

indirect anchor attribute relationship

Child popover

This is popover #1

Popover 2  Popover 4
This is popover #2

Popover 3
This is popover #3

This is popover #4

Popover 1 This is a dialogCloseDialog

FAIL Direct DOM children assert_false: expected false got true
FAIL Grandchildren assert_false: expected false got true
FAIL popovertoggletarget attribute relationship assert_true: expected true got false
FAIL nested popovertoggletarget attribute relationship assert_true: expected true got false
FAIL anchor attribute relationship assert_false: expected false got true
FAIL indirect anchor attribute relationship assert_false: expected false got true
FAIL more complex nesting, all using anchor ancestry assert_equals: Popover #2 incorrect state expected false but got true
FAIL popovers should be closed by dialogs Element has unexpected visibility state
FAIL dialogs should not be closed by popovers Element has unexpected visibility state
